refactor(car): outputConsoleView 주입
