+++ src/main/kotlin/racingcar/service/RacingCarService.kt
@@ -2,11 +2,9 @@ package racingcar.service
 
 import racingcar.domain.Cars
 import racingcar.domain.RandomCondition
-import racingcar.view.OutputView
+import racingcar.view.OutputConsoleView
 
-class RacingCarService {
-    private val outputView = OutputView()
-
+class RacingCarService(private val outputView: OutputConsoleView) {
     fun play(names: List<String>, attempts: Int): List<String> {
         val cars = Cars(names, RandomCondition())
