+++ src/main/kotlin/racingcar/view/InputConsoleView.kt
@@ -1,6 +1,6 @@
 package racingcar.view
 
-class InputView {
+class InputConsoleView {
     fun namesOfCars(): List<String> {
         println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
         return readln().split(",")
